requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.2
cssselect2==0.7.0
Django==4.2.3
django-phonenumber-field==7.1.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
oscrypto==1.3.0
pdfkit==1.0.0
phonenumbers==8.13.17
Pillow==10.0.0
pycparser==2.21
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pypdf==3.14.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.4
webencodings==0.5.1
xhtml2pdf==0.2.11
